retriever-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
elasticsearch==8.14.0
httpx==0.27.0
python-dotenv==1.0.1
